// src/app/shared/types/api-output-types/workscope-component/bire-item-for-workscope-tree-dto.interface.ts
export interface BireItemForWorkscopeTreeDTO {
  familyCode?: string;
  structureType?: string;
  variantCode?: string;
  chapter?: string;
  sheet?: string;
  marks?: string;
  subject?: string;
  section?: string;
  name?: string;
  itemType?: string;
  functionCode?: string;
  serializedPart?: boolean;
  itemFamilyCode?: string;
  itemNumber?: string;
  errorMessage?: string;
  icon?: string;
  isWorkscope?: boolean;
  item?: string;
  cause?: string;
  removalCause?: string;
  impact?: string;
  quantity?: string;
  expectedQuantity?: string;
  realQuantity?: string;
  resultAssignmentAlt?: string;
  resultAssignmentIcon?: string;
  isAdded?: boolean;
  isRemoved?: boolean;
  isModified?: boolean;
  inventoriesList?: unknown[];
  workscopeVersionNumber?: number;
  workscopeVersionStatus?: number;
}